Commands in background rework (#44)
* removed /commanddetached endpoints. The user must use /commands & /commandsfinished GET to retrieve the command information

+++ src/main/java/com/github/estuaryoss/agent/service/DbService.java
@@ -3,6 +3,7 @@ package com.github.estuaryoss.agent.service;
 import com.github.estuaryoss.agent.entity.ActiveCommand;
 import com.github.estuaryoss.agent.entity.FinishedCommand;
 import com.github.estuaryoss.agent.model.ProcessState;
+import com.github.estuaryoss.agent.model.api.CommandStatus;
 import com.github.estuaryoss.agent.repository.ActiveCommandRepository;
 import com.github.estuaryoss.agent.repository.FinishedCommandRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -11,8 +12,12 @@ import org.springframework.stereotype.Service;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
+import java.util.Comparator;
 import java.util.List;
 import java.util.stream.Collectors;
+
+import static com.github.estuaryoss.agent.constants.HibernateJpaConstants.*;
+import static com.github.estuaryoss.agent.utils.StringUtils.trimString;
 
 @Service
 public class DbService {
@@ -22,9 +27,12 @@ public class DbService {
     @Autowired
     private FinishedCommandRepository finishedCommandRepository;
 
-    public void save(String[] command, ProcessState processState) {
+    private final Comparator<FinishedCommand> finishedCommandComparator = Comparator.comparing(finishedCommand -> finishedCommand.getFinishedAt());
+
+    public void saveActiveCommand(String command, String commandId, ProcessState processState) {
         ActiveCommand activeCommand = ActiveCommand.builder()
-                .command(joinCommand(command))
+                .commandId(trimString(commandId, FIELD_MAX_SIZE))
+                .command(trimString(command, COMMAND_MAX_SIZE))
                 .startedAt(LocalDateTime.now().format(DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm:ss.SSSSSS")))
                 .pid(processState.getProcess().pid())
                 .build();
@@ -32,13 +40,24 @@ public class DbService {
         activeCommandRepository.saveAndFlush(activeCommand);
     }
 
-    public void clearAll() {
-        activeCommandRepository.deleteAll();
+    public void saveFinishedCommand(String command, String commandId, CommandStatus commandStatus) {
+        FinishedCommand finishedCommand = FinishedCommand.builder()
+                .commandId(trimString(commandId, FIELD_MAX_SIZE))
+                .command(trimString(command, COMMAND_MAX_SIZE))
+                .code(commandStatus.getDetails().getCode())
+                .out(trimString(commandStatus.getDetails().getOut(), COMMAND_STDOUT_MAX_SIZE))
+                .err(trimString(commandStatus.getDetails().getErr(), COMMAND_STDERR_MAX_SIZE))
+                .startedAt(commandStatus.getStartedat())
+                .finishedAt(commandStatus.getFinishedat())
+                .duration(commandStatus.getDuration())
+                .pid(commandStatus.getDetails().getPid())
+                .build();
+
+        finishedCommandRepository.saveAndFlush(finishedCommand);
     }
 
-    public List<ActiveCommand> getActiveCommands() {
-        return activeCommandRepository.findAll(Sort.by(Sort.Order.desc("id")))
-                .stream().collect(Collectors.toList());
+    public void clearAllActiveCommands() {
+        activeCommandRepository.deleteAll();
     }
 
     public List<FinishedCommand> getFinishedCommands(long limit) {
@@ -47,10 +66,25 @@ public class DbService {
                 .stream().limit(limit).collect(Collectors.toList());
     }
 
-    public void remove(long pid) {
-        ActiveCommand activeCommand = activeCommandRepository.findActiveCommandByPid(pid);
+    public List<ActiveCommand> getAllActiveCommands() {
+        return activeCommandRepository.findAll(Sort.by(Sort.Order.desc("id")))
+                .stream().collect(Collectors.toList());
+    }
 
-        activeCommandRepository.delete(activeCommand);
+    public void removeActiveCommand(long pid) {
+        ActiveCommand activeCommand = activeCommandRepository.findActiveCommandByPid(pid);
+        if (activeCommand != null) activeCommandRepository.delete(activeCommand);
+    }
+
+    public List<ActiveCommand> getAllActiveCommandsByCommandId(String commandId) {
+        return activeCommandRepository.findActiveCommandByCommandId(commandId);
+    }
+
+    public List<FinishedCommand> getAllFinishedCommandsByCommandId(String commandId) {
+        List<FinishedCommand> finishedCommandsList = finishedCommandRepository.findFinishedCommandByCommandId(commandId);
+        finishedCommandsList.sort(finishedCommandComparator);
+
+        return finishedCommandsList;
     }
 
     private String joinCommand(String[] command) {
